# generated-client/models/paginated-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Privet API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Paginated response schema.
 * @export
 * @interface PaginatedResponse
 */
export interface PaginatedResponse {
    /**
     * List of items
     * @type {Array<any>}
     * @memberof PaginatedResponse
     */
    'items': Array<any>;
    /**
     * Total number of items
     * @type {number}
     * @memberof PaginatedResponse
     */
    'total': number;
    /**
     * Current offset
     * @type {number}
     * @memberof PaginatedResponse
     */
    'offset': number;
    /**
     * Current limit
     * @type {number}
     * @memberof PaginatedResponse
     */
    'limit': number;
}