# generated-client/models/user-create.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Privet API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Schema for creating a user.
 * @export
 * @interface UserCreate
 */
export interface UserCreate {
    /**
     * 
     * @type {string}
     * @memberof UserCreate
     */
    'username'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserCreate
     */
    'first_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserCreate
     */
    'last_name'?: string | null;
    /**
     * User language code
     * @type {string}
     * @memberof UserCreate
     */
    'language'?: string;
    /**
     * Telegram user ID
     * @type {number}
     * @memberof UserCreate
     */
    'user_id': number;
}